pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "photo-organizer"
version = "0.1.0"
description = "Utilities for organizing photos locally"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "Pillow",
    "mediapipe",
    "onnxruntime",
    "numpy",
    "scikit-learn",
    "torch",
    "torchvision",
    "pytesseract",
    "geopy",
]

[project.scripts]
photo-organizer = "cli:main"

[tool.setuptools]
packages = ["photo_organizer"]
py-modules = ["cli"]
